# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Snake & Ladder Game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GameInput {
  hostName: string;
  level: number;
}

export interface JoinInput {
  playerName: string;
}

export interface JoinResult {
  playerId: string;
  roomCode: string;
}

export interface Player {
  id: string;
  name: string;
  position: number;
  score: number;
  color: string;
  bonusRolls?: number;
}

export interface Game {
  roomCode: string;
  level: number;
  status: string;
  players: Player[];
  currentPlayerIndex: number;
  /** @nullable */
  winnerId?: string | null;
}

export interface ErrorResponse {
  error: string;
}