# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Snake & Ladder Game API
 * OpenAPI spec version: 0.1.0
 */
import type { Player } from './player';

export interface Game {
  roomCode: string;
  level: number;
  status: string;
  players: Player[];
  currentPlayerIndex: number;
  /** @nullable */
  winnerId?: string | null;
}