# lib/api-zod/src/generated/types/gameInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Snake & Ladder Game API
 * OpenAPI spec version: 0.1.0
 */

export interface GameInput {
  hostName: string;
  level: number;
}