# lib/api-zod/src/generated/types/joinResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Snake & Ladder Game API
 * OpenAPI spec version: 0.1.0
 */

export interface JoinResult {
  playerId: string;
  roomCode: string;
}