# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Snake & Ladder Game API
 * OpenAPI spec version: 0.1.0
 */

export interface Player {
  id: string;
  name: string;
  position: number;
  score: number;
  color: string;
  bonusRolls?: number;
}